api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
supabase==2.0.1
google-generativeai==0.7.2
serpapi==0.1.3
pydantic==2.10.0
httpx>=0.24.0,<0.25.0
python-multipart==0.0.18